Change the data type of Temperature to float

+++ src/main/java/org/conserve/Rooms.java
@@ -15,7 +15,7 @@ abstract class Rooms {
         this.roomID = roomID;
     }
 
-    public float getTemperature(){
+    public double getTemperature(){
         return temperature;
     }
